http*: add helper methods for fetching objects/info/packs
The code handling the fetching of objects/info/packs in http-push.c and
http-walker.c have been refactored into new methods and a new struct
in http.c. They are not meant to be invoked elsewhere.

The new methods in http.c are
 *new_http_info_packs_request
 *finish_http_info_packs_request
 *release_http_info_packs_request

and the new struct is http_info_packs_request.

A callback function wraps the invocation of fetch_http_pack_index,
invoked in finish_http_info_packs_request when parsing info/packs.

The fetching code lives in three methods to accomodate http-walker.c's
implementation, which involves the variable repo->got_indices.