transport-helper: abort gracefully
Abort gracefully in abort_helper() using the same code as
disconnect_helper(), sans sending a '\n' to the helper's stdin and
invoking finish_command() on the helper process.

This allows the transport layer and higher (eg. git-push) to alert the
user to any errors reported by the remote helper.

Previously, the transport-helper layer just did a exit().

Signed-off-by: Tay Ray Chuan <rctay89@gmail.com>