transport-helper::push_refs: mark refs to be pushed with 'expecting report'
Mark refs that are to be pushed (ie. sent to the helper with a "push"
line) with status REF_STATUS_EXPECTING_REPORT.

Currently, refs to be pushed have the status REF_STATUS_NONE. However,
the implicit use of this status as "ref should be pushed" is later
conflated with 'no match' when transport users analyse ref statuses.

For example, if a remote helper failed to give a status report for a ref
(ie. "ok" or "error" line), transport users took it as though the ref
was not pushed in a non-errorneous fashion (like REF_STATUS_UPTODATE).

Marking refs to be pushed as 'expecting report' removes this ambiguity,
and users can be clued in to any errors that might have ocurred (eg. in
the remote helper during pushing) via print_push_status().

Signed-off-by: Tay Ray Chuan <rctay89@gmail.com>