commit::print_summary(): set rev_info.always_show_header to 1
This attempts to fix a regression in git-commit, where non-abbreviated
SHA-1s were printed in the summary.

One possible fix would be to set ctx.abbrev to DEFAULT_ABBREV in the
`if` block.

However, we remove this codeblock altogether, and set
rev.always_show_header. This way, we use back the same show_log()
mechanism (instead of format_commit_message()).

Quoting log-tree.c:560:

	shown = log_tree_diff(opt, commit, &log);
	if (!shown && opt->loginfo && opt->always_show_header) {
		log.parent = NULL;
		show_log(opt);
		shown = 1;
	}

This is the only area that always_show_header is checked, so the
setting of this flag should only affect this area.

Also, we now die() if log_tree_commit() returns false. Based on the
existing underlying codepaths (log_tree_commit(), log_tree_diff(),
log_tree_diff_flush(), to name a few), this should not occur; changes to
these codepaths may warrant a revision of our handling of this
situation. Tests #2 and #3 in t7502 should aid in detecting such
breakages.

Signed-off-by: Tay Ray Chuan <rctay89@gmail.com>